export decode/encode from meta store
