+++ src/blockstore/index.ts
@@ -14,3 +14,4 @@ export { EncryptedBlockstore, BaseBlockstore, CompactionFetcher, type BlockFetch
 export { Loader } from "./loader.js";
 export { parseCarFile } from "./loader-helpers.js";
 export { ConnectionBase } from "./connection-base.js";
+export { decodeGatewayMetaBytesToDbMeta, encodeGatewayDbMetaToBytes } from "./store-meta.js";
