+++ src/blockstore/store-remote.ts
@@ -1,5 +1,6 @@
 import { URI } from "@adviser/cement";
-import { DataStoreImpl, MetaStoreImpl, StoreOpts } from "./store.js";
+import { DataStoreImpl, StoreOpts } from "./store.js";
+import { MetaStoreImpl } from "./store-meta.js";
 import { SuperThis } from "../types.js";
 
 // export type LoadHandler = (dbMetas: DbMeta[]) => Promise<void>;
