+++ src/blockstore/store.ts
@@ -1,9 +1,6 @@
 import pLimit from "p-limit";
 import { format, parse, ToString } from "@ipld/dag-json";
 import { exception2Result, Logger, ResolveOnce, Result, URI } from "@adviser/cement";
-import { EventBlock, decodeEventBlock } from "@web3-storage/pail/clock";
-import { EventView } from "@web3-storage/pail/clock/api";
-
 import type {
   AnyBlock,
   AnyLink,
@@ -11,17 +8,12 @@ import type {
   DataSaveOpts,
   DataStore,
   DbMeta,
-  MetaStore,
   WALStore as WALStore,
   WALState,
-  LoadHandler,
   KeyedCrypto,
   Loadable,
-  CarClockHead,
-  DbMetaEventBlock,
-  CarClockLink,
 } from "./types.js";
-import { Falsy, StoreType, SuperThis, throwFalsy, CRDTEntry } from "../types.js";
+import { Falsy, StoreType, SuperThis, throwFalsy } from "../types.js";
 import { Gateway } from "./gateway.js";
 import { ensureLogger, isNotFoundError } from "../utils.js";
 import { carLogIncludesGroup } from "./loader.js";
@@ -29,7 +21,6 @@ import { CommitQueue } from "./commit-queue.js";
 import { keyedCryptoFactory } from "../runtime/keyed-crypto.js";
 import { KeyBag } from "../runtime/key-bag.js";
 import { FragmentGateway } from "./fragment-gateway.js";
-import { Link } from "multiformats";
 
 function guardVersion(url: URI): Result<URI> {
   if (!url.hasParam("version")) {
@@ -44,7 +35,7 @@ export interface StoreOpts {
   readonly loader?: Loadable;
 }
 
-abstract class BaseStoreImpl {
+export abstract class BaseStoreImpl {
   // should be injectable
 
   abstract readonly storeType: StoreType;
@@ -131,137 +122,6 @@ abstract class BaseStoreImpl {
     this._onStarted.forEach((fn) => fn());
     this.logger.Debug().Msg("started");
     return version;
-  }
-}
-
-export class MetaStoreImpl extends BaseStoreImpl implements MetaStore {
-  readonly storeType = "meta";
-  readonly subscribers = new Map<string, LoadHandler[]>();
-  parents: CarClockHead = [];
-
-  constructor(sthis: SuperThis, name: string, url: URI, opts: StoreOpts, remote?: boolean) {
-    // const my = new URL(url.toString());
-    // my.searchParams.set("storekey", 'insecure');
-    super(
-      name,
-      url,
-      {
-        ...opts,
-      },
-      sthis,
-      ensureLogger(sthis, "MetaStoreImpl"),
-    );
-    if (remote && opts.gateway.subscribe) {
-      this.logger.Debug().Str("url", url.toString()).Msg("Subscribing to the gateway");
-      opts.gateway.subscribe(url, (byteHead: Uint8Array) => this.handleEventByteHead(byteHead));
-    }
-  }
-
-  makeHeader({ cars }: DbMeta): ToString<DbMeta> {
-    const toEncode: DbMeta = { cars };
-    // if (key) toEncode.key = key;
-    return format(toEncode);
-  }
-
-  async createEventBlock(bytes: Uint8Array): Promise<DbMetaEventBlock> {
-    const data = {
-      dbMeta: bytes,
-    };
-    const event = await EventBlock.create(
-      data,
-      this.parents as unknown as Link<EventView<{ dbMeta: Uint8Array }>, number, number, 1>[],
-    );
-    // await this.eventBlocks.put(event.cid, event.bytes);
-    return event as EventBlock<{ dbMeta: Uint8Array }>;
-  }
-
-  async decodeEventBlock(bytes: Uint8Array): Promise<DbMetaEventBlock> {
-    const event = await decodeEventBlock<{ dbMeta: Uint8Array }>(bytes);
-    return event as EventBlock<{ dbMeta: Uint8Array }>;
-  }
-
-  async decodeMetaBlock(crdtEntry: CRDTEntry): Promise<{ eventCid: CarClockLink; dbMeta: DbMeta; parents: string[] }> {
-    const eventBytes = decodeFromBase64(crdtEntry.data);
-    const eventBlock = await this.decodeEventBlock(eventBytes);
-    return {
-      eventCid: eventBlock.cid as CarClockLink,
-      parents: crdtEntry.parents,
-      dbMeta: parse<DbMeta>(this.sthis.txt.decode(eventBlock.value.data.dbMeta)),
-    };
-  }
-
-  async handleByteHeads(byteHeads: Uint8Array) {
-    const crdtEntries = JSON.parse(this.sthis.txt.decode(byteHeads)) as CRDTEntry[];
-    try {
-      const dbMetas = await Promise.all(crdtEntries.map((entry) => this.decodeMetaBlock(entry)));
-      return dbMetas;
-    } catch (e) {
-      throw this.logger.Error().Err(e).Msg("parseHeader").AsError();
-    }
-  }
-
-  async handleEventByteHead(byteHead: Uint8Array) {
-    const crdtEntry = JSON.parse(this.sthis.txt.decode(byteHead)) as CRDTEntry;
-    const { eventCid, dbMeta, parents } = await this.decodeMetaBlock(crdtEntry);
-    this.loader?.taskManager?.handleEvent(eventCid, parents, dbMeta);
-  }
-
-  async load(): Promise<DbMeta[] | Falsy> {
-    const branch = "main";
-    const url = await this.gateway.buildUrl(this.url(), branch);
-    if (url.isErr()) {
-      throw this.logger.Error().Result("buidUrl", url).Str("branch", branch).Msg("got error from gateway.buildUrl").AsError();
-    }
-    const bytes = await this.gateway.get(url.Ok());
-    if (bytes.isErr()) {
-      if (isNotFoundError(bytes)) {
-        return undefined;
-      }
-      throw this.logger.Error().Url(url.Ok()).Result("bytes:", bytes).Msg("gateway get").AsError();
-    }
-    const dbMetas = await this.handleByteHeads(bytes.Ok());
-    await this.loader?.handleDbMetasFromStore(dbMetas.map((m) => m.dbMeta)); // the old one didn't await
-    const cids = dbMetas.map((m) => m.eventCid);
-    const uniqueParentsMap = new Map([...this.parents, ...cids].map((p) => [p.toString(), p]));
-    this.parents = Array.from(uniqueParentsMap.values());
-    return dbMetas.map((m) => m.dbMeta);
-  }
-
-  async encodeEventWithParents(event: EventBlock<{ dbMeta: Uint8Array }>): Promise<Uint8Array> {
-    const base64String = encodeToBase64(event.bytes);
-    const crdtEntry = {
-      cid: event.cid.toString(),
-      data: base64String,
-      parents: this.parents.map((p) => p.toString()),
-    };
-    return this.sthis.txt.encode(JSON.stringify([crdtEntry]));
-  }
-
-  async save(meta: DbMeta, branch?: string): Promise<Result<void>> {
-    branch = branch || "main";
-    this.logger.Debug().Str("branch", branch).Any("meta", meta).Msg("saving meta");
-    const event = await this.createEventBlock(this.sthis.txt.encode(this.makeHeader(meta)));
-    const bytes = await this.encodeEventWithParents(event);
-    const url = await this.gateway.buildUrl(this.url(), branch);
-    if (url.isErr()) {
-      throw this.logger.Error().Err(url.Err()).Str("branch", branch).Msg("got error from gateway.buildUrl").AsError();
-    }
-    const res = await this.gateway.put(url.Ok(), bytes);
-    if (res.isErr()) {
-      throw this.logger.Error().Err(res.Err()).Msg("got error from gateway.put").AsError();
-    }
-    await this.loader?.handleDbMetasFromStore([meta]);
-    this.parents = [event.cid];
-    return res;
-  }
-
-  async close(): Promise<Result<void>> {
-    await this.gateway.close(this.url());
-    this._onClosed.forEach((fn) => fn());
-    return Result.Ok(undefined);
-  }
-  async destroy(): Promise<Result<void>> {
-    return this.gateway.destroy(this.url());
   }
 }
 
@@ -522,49 +382,4 @@ export class WALStoreImpl extends BaseStoreImpl implements WALStore {
   destroy() {
     return this.gateway.destroy(this.url());
   }
-}
-
-function encodeToBase64(bytes: Uint8Array): string {
-  const chars = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
-  let base64 = "";
-  let i;
-  for (i = 0; i < bytes.length - 2; i += 3) {
-    base64 += chars[bytes[i] >> 2];
-    base64 += chars[((bytes[i] & 3) << 4) | (bytes[i + 1] >> 4)];
-    base64 += chars[((bytes[i + 1] & 15) << 2) | (bytes[i + 2] >> 6)];
-    base64 += chars[bytes[i + 2] & 63];
-  }
-  if (i < bytes.length) {
-    base64 += chars[bytes[i] >> 2];
-    if (i === bytes.length - 1) {
-      base64 += chars[(bytes[i] & 3) << 4];
-      base64 += "==";
-    } else {
-      base64 += chars[((bytes[i] & 3) << 4) | (bytes[i + 1] >> 4)];
-      base64 += chars[(bytes[i + 1] & 15) << 2];
-      base64 += "=";
-    }
-  }
-  return base64;
-}
-
-function decodeFromBase64(base64: string): Uint8Array {
-  const chars = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
-  const bytes = new Uint8Array((base64.length * 3) / 4);
-  let i;
-  let j = 0;
-  for (i = 0; i < base64.length; i += 4) {
-    const a = chars.indexOf(base64[i]);
-    const b = chars.indexOf(base64[i + 1]);
-    const c = chars.indexOf(base64[i + 2]);
-    const d = chars.indexOf(base64[i + 3]);
-    bytes[j++] = (a << 2) | (b >> 4);
-    if (base64[i + 2] !== "=") {
-      bytes[j++] = ((b & 15) << 4) | (c >> 2);
-    }
-    if (base64[i + 3] !== "=") {
-      bytes[j++] = ((c & 3) << 6) | d;
-    }
-  }
-  return bytes.slice(0, j);
 }
