+++ src/blockstore/types.ts
@@ -196,7 +196,7 @@ export interface CommitOpts {
 
 export interface DbMeta {
   readonly cars: CarGroup;
-  // key?: string;
+  key?: string;
 }
 
 // export interface UploadMetaFnParams {
