+++ src/types.ts
@@ -6,6 +6,8 @@ import { EnvFactoryOpts, Env, Logger, CryptoRuntime } from "@adviser/cement";
 
 import type { MakeDirectoryOptions, PathLike, Stats } from "fs";
 
+export type { DbMeta };
+
 export type Falsy = false | null | undefined;
 
 export function isFalsy(value: unknown): value is Falsy {
